Shopping cart layout almost complete
TODO: CSS when the layout is finialized.  Note: almost completed the
total price feature, just need to multiply the cost per unit by the
amount of units.  NOTE: 3 way data binding is adding a counting variable
to the firebase database.  This variable needs to remain server side so
as to not mess up the experience of multiple users.