Finished most styling for mobile modal L/R toggle buttons.
TODO: wire up the buttons, have new variables, an index number for each
item in each set of items.  New function increments number, or decreases
it depending on left or right click.  This index should be created on
page load and referenced as a local variable.  Could also have it stored
in firebase just in case user changes variables locally.