fix: addressing bootstrap modal issue for non-mobile
Bootstrap modal for desktop at 600px or less removes the scrollbar when
opened.  This has the undesired effect of breaking the code that
accounts for the R/L scrolling animation.  For chrome the pixel width
missing is 17px, which is the width of the missing scrollbar.

setting 'overflow: scroll' to the .modal-open css will fix this, however
we are left with two different scroll bars.  Will attempt to solve this
issue without altering code for R/L scroll widths.