fix: Correct hex color code and way of changing CSS
Hex color codes contained extra characters and thus were not being
rendered in the browser.  Also, instead of adding an inline style with
jQuery's .css() method, used .addClass() instead.  Allows for removal of
class when window is closed (or if want the notification to disappear,
blink, or do something else).